refactor: rename jupiterHelper to jupiterLendingHelper and clean up index.js
